requirements/prod.txt
Django==3.2.4
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
PyJWT==2.8.0
python-decouple==3.8